small refactor, place each functions into appropriate files for better clarity identifying each roles
