+++ tests/intent_test.go
@@ -9,7 +9,7 @@ import (
 func TestSwapIntentAbiEncode(t *testing.T) {
 	t.Parallel()
 
-	swapIntent := NewDemoSwapIntent()
+	swapIntent := newDemoSwapIntent()
 	want := common.FromHex("0x00000000000000000000000000000000000000000000000000000000000000200000000000000000000000007439e9bb6d8a84dd3a23fe621a30f95403f87fb90000000000000000000000000000000000000000000000000000b5e620f480000000000000000000000000007b79995e5f793a07bc00c21412e50ecae098e7f9000000000000000000000000000000000000000000000000000000e8d4a510000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000000")
 
 	result, err := swapIntent.abiEncode()
@@ -25,7 +25,7 @@ func TestSwapIntentAbiEncode(t *testing.T) {
 func TestSwapIntentAbiEncodeWithSelector(t *testing.T) {
 	t.Parallel()
 
-	swapIntent := NewDemoSwapIntent()
+	swapIntent := newDemoSwapIntent()
 	want := common.FromHex("0x83a6992a00000000000000000000000000000000000000000000000000000000000000200000000000000000000000007439e9bb6d8a84dd3a23fe621a30f95403f87fb90000000000000000000000000000000000000000000000000000b5e620f480000000000000000000000000007b79995e5f793a07bc00c21412e50ecae098e7f9000000000000000000000000000000000000000000000000000000e8d4a510000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000000")
 
 	result, err := swapIntent.abiEncodeWithSelector()
